Delete Group and fixed delete Folder

+++ app/src/main/java/com/example/piston/main/personal/folder/info/FolderInfoRepository.java
@@ -16,18 +16,17 @@ public class FolderInfoRepository {
     }
 
     private final FirebaseFirestore db = FirebaseFirestore.getInstance();
-
-    private final String folderID, user;
+    private final DocumentReference docRef;
 
     public FolderInfoRepository(FolderInfoRepository.IFolderInfo listener, String folderID) {
-        this.folderID = folderID;
-        user = Objects.requireNonNull(FirebaseAuth.getInstance().getCurrentUser()).getEmail();
+        String user = Objects.requireNonNull(FirebaseAuth.getInstance().getCurrentUser()).getEmail();
 
-        db.collection("users")
+        docRef = db.collection("users")
                 .document(Objects.requireNonNull(user))
                 .collection("folders")
-                .document(folderID)
-                .get().addOnCompleteListener(task -> {
+                .document(folderID);
+
+        docRef.get().addOnCompleteListener(task -> {
             if (task.isSuccessful()) {
                 listener.setTitle((String) Objects.requireNonNull(task.getResult()).get("title"));
                 listener.setDescription((String) task.getResult().get("description"));
@@ -36,22 +35,30 @@ public class FolderInfoRepository {
     }
 
     public void deleteFolder() {
-        DocumentReference docRef = db.collection("users")
-                .document(user)
-                .collection("folders")
-                .document(folderID);
-
         docRef.collection("posts").get().addOnCompleteListener(task -> {
             if (task.isComplete()) {
                 // Delete subcollection
-                for (QueryDocumentSnapshot documentSnapshot : Objects.requireNonNull(
+                for (QueryDocumentSnapshot snapshot : Objects.requireNonNull(
                         task.getResult())) {
-                    docRef.collection("posts")
-                            .document(documentSnapshot.getId())
-                            .delete();
+                    DocumentReference docRef1 = docRef.collection("posts")
+                            .document(snapshot.getId());
+
+                    docRef1.collection("replies")
+                            .get().addOnCompleteListener(task1 -> {
+                                if (task1.isSuccessful()) {
+                                    for (QueryDocumentSnapshot snapshot1 : Objects.requireNonNull(
+                                            task1.getResult())) {
+                                        docRef1.collection("replies")
+                                                .document(snapshot1.getId())
+                                                .delete();
+                                    }
+                                }
+                    });
+                    docRef1.delete();
                 }
             }
-            docRef.delete();
         });
+        docRef.delete();
     }
+
 }
